pyproject.toml
[project]
name = "ai-meeting-notes-jira-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.16",
    "fastapi>=0.127.1",
    "ffmpeg-python>=0.2.0",
    "openai>=2.14.0",
    "openai-whisper>=20250625",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "streamlit>=1.52.2",
    "streamlit-autorefresh>=1.0.1",
    "uvicorn[standard]>=0.40.0",
    "websocket-client>=1.9.0",
    "yt-dlp>=2025.12.8",
]
